+++ NEWS.md
@@ -1,3 +1,10 @@
+# socialranking 0.1.1
+
+* Added parameter `equivalenceClasses` to `newPowerRelation()`
+* Removed kableExtra from Suggests
+* Fix `equivalenceClassIndex()` takes cycles into account
+* Fix lexcel equals operator
+
 # socialranking 0.1.0 (2022-03-28)
 
 * Initial release